% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_pff_experiments.R
\name{are_pff_experiments}
\alias{are_pff_experiments}
\title{Determine if all experiments have Peregrine friendly filenames}
\usage{
are_pff_experiments(experiments)
}
\arguments{
\item{experiments}{a list of \code{pirouette} experiments,
as created by \link[pirouette]{create_experiment}}
}
\description{
Determine if all experiments have Peregrine friendly filenames
}
