% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pff_alignment_params.R
\name{check_pff_alignment_params}
\alias{check_pff_alignment_params}
\title{Check if the alignment parameter have PFF}
\usage{
check_pff_alignment_params(alignment_params)
}
\arguments{
\item{alignment_params}{parameters for creating an alignment,
as can be created by \code{\link[pirouette]{create_alignment_params}}}
}
\description{
Check if the alignment parameter have PFF
}
