% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pff_beast2_options.R
\name{check_pff_beast2_options}
\alias{check_pff_beast2_options}
\title{Check if the BEAST2 options are PFF}
\usage{
check_pff_beast2_options(beast2_options)
}
\arguments{
\item{beast2_options}{BEAST2 options,
as can be created by \link[beastier]{create_beast2_options}}
}
\description{
Check if the BEAST2 options are PFF
}
