% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pff_experiment.R
\name{check_pff_experiment}
\alias{check_pff_experiment}
\title{Check if an experiment has PFF}
\usage{
check_pff_experiment(experiment)
}
\arguments{
\item{experiment}{a \link{pirouette} experiment,
as can be created by \link{create_experiment}}
}
\description{
Check if an experiment has PFF
}
