% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pff_experiments.R
\name{check_pff_experiments}
\alias{check_pff_experiments}
\title{Check if all experiments have PFF}
\usage{
check_pff_experiments(experiments)
}
\arguments{
\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}. If more than one experiment
is provided and a "generative" experiment is part of them, the "generative"
one has to be the first in the list.}
}
\description{
Check if all experiments have PFF
}
