% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pff_inference_model.R
\name{check_pff_inference_model}
\alias{check_pff_inference_model}
\title{Check if the inference model is Peregrine friendly}
\usage{
check_pff_inference_model(inference_model)
}
\description{
Will \link{stop} if not
}
