% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pff_mcmc.R
\name{check_pff_mcmc}
\alias{check_pff_mcmc}
\title{Check if the mcmc is PFF}
\usage{
check_pff_mcmc(mcmc)
}
\arguments{
\item{mcmc}{an MCMC,
as created by \link[beautier]{create_mcmc} or by
\link{create_pff_mcmc} for automatic peregrine friendly filenames check}
}
\description{
Check if the mcmc is PFF
}
