% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pff_pir_params.R
\name{check_pff_pir_params}
\alias{check_pff_pir_params}
\title{Check if the \code{pir_params} are valid and Peregrine-friendly.}
\usage{
check_pff_pir_params(pir_params)
}
\arguments{
\item{pir_params}{\code{pirouette} parameters,
as created by \link[pirouette]{create_pir_params}}
}
\description{
Will \link{stop} if the \code{pir_params} are invalid
or Peregrine-unfriendly, else will do nothing
}
\author{
Richel J.C. Bilderbeek
}
