% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pff_twinning_params.R
\name{check_pff_twinning_params}
\alias{check_pff_twinning_params}
\title{Check if the twinning parameters have PFF}
\usage{
check_pff_twinning_params(twinning_params)
}
\arguments{
\item{twinning_params}{parameters for creating a twin tree,
as can be created by \code{\link[pirouette]{create_twinning_params}}}
}
\description{
Check if the twinning parameters have PFF
}
