% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_regular_jobs.R
\name{count_regular_jobs}
\alias{count_regular_jobs}
\title{Count the number of \code{regular} jobs}
\usage{
count_regular_jobs(user_name = "$USER")
}
\arguments{
\item{user_name}{the Peregrine username, for example \code{p123456}}
}
\value{
the number of jobs
}
\description{
Count the number of \code{regular} jobs that are running on Peregrine for
a specific user.
}
\details{
Use \link{count_jobs} to count all jobs running.
Use \link{count_vulture_jobs} to count the jobs running
on the \code{culture} partition.
}
\examples{
if (is_on_peregrine()) {
  count_regular_jobs()
}
}
