% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_vulture_jobs.R
\name{count_vulture_jobs}
\alias{count_vulture_jobs}
\title{Count the number of \code{vulture} jobs}
\usage{
count_vulture_jobs(user_name = "$USER")
}
\arguments{
\item{user_name}{the Peregrine username, for example \code{p123456}}
}
\value{
the number of jobs
}
\description{
Count the number of all jobs that are running on Peregrine for
a specific user.
}
\details{
Use \link{count_jobs} to count all jobs running
on the \code{regular} partition.
Use \link{count_regular_jobs} to count the jobs running
on the \code{regular} partition.
}
\examples{
if (is_on_peregrine()) {
  count_vulture_jobs()
}
}
