% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_all_pff_experiments.R
\name{create_all_pff_experiments}
\alias{create_all_pff_experiments}
\title{Create all experiments as done by
\link[pirouette]{create_all_experiments}
with PFF}
\usage{
create_all_pff_experiments(site_models = beautier::create_site_models(),
  clock_models = beautier::create_clock_models(),
  tree_priors = beautier::create_tree_priors(),
  mcmc = beautier::create_mcmc(store_every = 1000), exclude_model = NA)
}
\description{
Create all experiments as done by
\link[pirouette]{create_all_experiments}
with PFF
}
