% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pff_mcmc.R
\name{create_pff_mcmc}
\alias{create_pff_mcmc}
\title{Create mcmc that has Peregrine friendly filenames}
\usage{
create_pff_mcmc(chain_length = beautier::create_mcmc()$chain_length,
  store_every = beautier::create_mcmc()$store_every,
  pre_burnin = beautier::create_mcmc()$pre_burnin,
  n_init_attempts = beautier::create_mcmc()$n_init_attempts,
  sample_from_prior = beautier::create_mcmc()$sample_from_prior,
  tracelog = beautier::create_tracelog(filename =
  peregrine::get_pff_tempfile()),
  screenlog = beautier::create_screenlog(filename = ""),
  treelog = beautier::create_treelog(filename =
  peregrine::get_pff_tempfile()))
}
\arguments{
\item{chain_length}{something}
}
\description{
Create mcmc that has Peregrine friendly filenames
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
