% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pff_twinning_params.R
\name{create_pff_twinning_params}
\alias{create_pff_twinning_params}
\title{Create a PFF twinning parameter set}
\usage{

  create_pff_twinning_params(rng_seed_twin_tree = pirouette::create_twinning_params()$rng_seed_twin_tree,
  sim_twin_tree_fun = pirouette::create_twinning_params()$sim_twin_tree_fun,
  rng_seed_twin_alignment = pirouette::create_twinning_params()$rng_seed_twin_alignment,
  sim_twin_alignment_fun = pirouette::create_twinning_params()$sim_twin_alignment_fun,
  twin_tree_filename = peregrine::get_pff_tempfile(pattern =
  "tree_twin_", fileext = ".newick"),
  twin_alignment_filename = peregrine::get_pff_tempfile(pattern =
  "alignment_twin_", fileext = ".fasta"),
  twin_evidence_filename = peregrine::get_pff_tempfile(pattern =
  "evidence_twin_", fileext = ".csv"))
}
\arguments{
\item{rng_seed_twin_tree}{the random number generator seed as used in the
simulation of a twin tree}

\item{rng_seed_twin_alignment}{the random number generator seed
as used in the simulation of a twin alignment}

\item{twin_tree_filename}{name of the (\code{.newick}) file the twin
tree will be saved to}

\item{twin_alignment_filename}{name of the FASTA file the twin
alignment will be saved to}

\item{twin_evidence_filename}{filename to store the estimated
evidences (aka marginal likelihoods) of the twin tree}
}
\description{
Create a PFF twinning parameter set
}
