% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_pff_pir_params.R
\name{create_test_pff_pir_params}
\alias{create_test_pff_pir_params}
\title{Create a \code{pir_params} with PFFs}
\usage{
create_test_pff_pir_params(
  alignment_params = pirouette::create_alignment_params(
    fasta_filename = get_pff_tempfile()
  ),
  experiments = list(create_test_pff_gen_experiment())
)
}
\arguments{
\item{alignment_params}{parameters for creating an alignment,
as can be created by \code{\link[pirouette]{create_alignment_params}}}

\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}. If more than one experiment
is provided and a "generative" experiment is part of them, the "generative"
one has to be the first in the list.}
}
\description{
Create a \code{pir_params} with PFFs
}
\author{
Richel J.C. Bilderbeek
}
