% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(alignment, alignment_params, bd_mutation_rate, bd_tree,
  bd_tree_filename, beast2_options, beast2_working_dir, brts, chain_length,
  clock_model, clock_models, cond, crown_age, error_measure_params,
  exclude_model, experiment, experiments, experiment_type, fasta_filename,
  fileext, filename, folder_name, init_speciation_rate,
  init_extinction_rate, input_filename, lambda, mbd_l_matrix,
  mbd_mutation_rate, mbd_params, mbd_paramses, mbd_params_interval,
  mbd_sim_rng_seed, mbd_tree, mcmc, mcmc_chain_length, misc_params,
  mrca_prior, mu, mutation_rate, n_replicates, nu, nu_events,
  output_log_filename, output_trees_filenames, output_state_filename,
  parameter_filename, parameters_filename, pattern, pff_tmpdir, phylo,
  pir_params, posterior_trees, precision, project_folder_name, q,
  sample_interval, seed, sequence_length, sim_pars, sim_phylo, site_model,
  site_models, sub_chain_length, tree, tree_prior, tree_priors,
  tree_filename, trees_filename, twinning_params, verbose)
}
\arguments{
\item{alignment}{a DNA alignment}

\item{alignment_params}{parameters for creating an alignment,
as can be created by \code{\link[pirouette]{create_alignment_params}}}

\item{bd_mutation_rate}{the mutation rate when creating an alignment
from a BD tree}

\item{bd_tree}{a phylogent of class \code{phylo}, created by a Birth Death
process}

\item{bd_tree_filename}{name of the file that stores a BD twin tree}

\item{beast2_options}{BEAST2 options,
as can be created by \link[beastier]{create_beast2_options}}

\item{beast2_working_dir}{BEAST2 temporary working directory,
as used in \link[beastier]{create_beast2_options}}

\item{brts}{set of branching times}

\item{chain_length}{something}

\item{clock_model}{a clock model,
as created by \link[beautier]{create_clock_model}}

\item{clock_models}{a list of one or more clock models,
as created by \link[beautier]{create_clock_model}}

\item{cond}{conditioning as specified in \code{\link[mbd]{mbd_sim}}}

\item{crown_age}{The crown age of the tree.}

\item{error_measure_params}{parameters to set how the error
between given tree and inferred trees in measure,
as can be created by \code{\link[pirouette]{create_error_measure_params}}}

\item{exclude_model}{an inference model that has to be excluded, as can be
created by \link[beautier]{create_inference_model}}

\item{experiment}{a \link{pirouette} experiment,
as can be created by \link{create_experiment}}

\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}. If more than one experiment
is provided and a "generative" experiment is part of them, the "generative"
one has to be the first in the list.}

\item{experiment_type}{the type of experiment,
can be either \code{test} or \code{full}}

\item{fasta_filename}{name of a FASTA file}

\item{fileext}{see \link{tempfile}}

\item{filename}{the file's name, without the path}

\item{folder_name}{name of the main folder}

\item{init_speciation_rate}{a speciation rate}

\item{init_extinction_rate}{an extinction rate}

\item{input_filename}{BEAST2 input file name,
as used in \link[beastier]{create_beast2_options}}

\item{lambda}{per-lineage speciation rate. See \code{\link[mbd]{mbd_sim}}}

\item{mbd_l_matrix}{the L matrix of an MBD tree}

\item{mbd_mutation_rate}{the mutation rate when creating an alignment
from a MBD tree}

\item{mbd_params}{MBD parameter set}

\item{mbd_paramses}{data frame with one MBD parameter set per row.
If a simulation with one MBD parameter set has 10 replicates,
this data frame will hold 10 rows, each with different seeds}

\item{mbd_params_interval}{data frame with one MBD parameter set per row.
If a simulation with one MBD parameter set has 10 replicates,
this data frame will hold 10 rows, each with different seeds}

\item{mbd_sim_rng_seed}{rng seed to simulate a mbd tree}

\item{mbd_tree}{an MBD tree}

\item{mcmc}{an MCMC,
as created by \link[beautier]{create_mcmc}}

\item{mcmc_chain_length}{length of an MCMC}

\item{misc_params}{additional parameters for razzo. They contain
tree_filename to store the original given tree and mbd_sim_rng_seed for
when an mbd tree is simulated}

\item{mrca_prior}{an MRCA priors,
as created by \link[beautier]{create_mrca_prior}}

\item{mu}{per-species extinction rate. See \code{\link[mbd]{mbd_sim}}}

\item{mutation_rate}{something}

\item{n_replicates}{number of replicates}

\item{nu}{the rate at which a multiple-birth specation is triggered.
See \code{\link[mbd]{mbd_sim}}}

\item{nu_events}{the number of nu-triggered events that have to be
present in the simulated tree}

\item{output_log_filename}{BEAST2 ouput \code{.log} file name,
as used in \link[beastier]{create_beast2_options}}

\item{output_trees_filenames}{BEAST2 ouput \code{.trees} file name,
as used in \link[beastier]{create_beast2_options}}

\item{output_state_filename}{BEAST2 ouput \code{.xml.state} file name,
as used in \link[beastier]{create_beast2_options}}

\item{parameter_filename}{full path to a \code{parameters.RDa} file}

\item{parameters_filename}{full path to a \code{parameters.RDa} file}

\item{pattern}{see \link{tempfile}}

\item{pff_tmpdir}{Peregrine-friendly temporary directory name,
see \link{get_pff_tempdir}}

\item{phylo}{a phylogeny}

\item{pir_params}{\code{pirouette} parameters,
as created by \link[pirouette]{create_pir_params}}

\item{posterior_trees}{phylogenetic trees in a BEAST2 posterior,
of class \code{multiphylo}}

\item{precision}{define the precision of the approximation.}

\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}

\item{q}{per-species speciation probability in case of occurrance of
a multiple event. See \code{\link[mbd]{mbd_sim}}}

\item{sample_interval}{the interval at which the MCMC algorithm
makes a measurement}

\item{seed}{a random number generator seed}

\item{sequence_length}{the length of each DNA sequence in an alignment}

\item{sim_pars}{something}

\item{sim_phylo}{something}

\item{site_model}{a nucleotide substitution model,
as created by \link[beautier]{create_site_model}}

\item{site_models}{a list of one or more site models,
as created by \link[beautier]{create_site_model}}

\item{sub_chain_length}{length of the sub-chain used by the Nested Sampling
algorithm to estimate the marginal likelihood}

\item{tree}{an ultrametric phylogenetic tree of class \code{phylo}}

\item{tree_prior}{a tree prior,
as created by \link[beautier]{create_tree_prior}}

\item{tree_priors}{a list of one or more tree priors,
as created by \link[beautier]{create_tree_prior}}

\item{tree_filename}{name of the phylogeny file}

\item{trees_filename}{name of the BEAST2 posterior phylogenies file}

\item{twinning_params}{parameters for creating a twin tree,
as can be created by \code{\link[pirouette]{create_twinning_params}}}

\item{verbose}{give more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno,
  use of this function by Richel J.C. Bilderbeek
}
