% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pff_tempdir.R
\name{get_pff_tempdir}
\alias{get_pff_tempdir}
\title{Get a Peregrine-friendly temporary directory}
\usage{
get_pff_tempdir()
}
\description{
Get a Peregrine-friendly temporary directory
}
\examples{
is_pff(get_pff_tempdir())
is_pff(get_pff_tempfile())
is_pff("/peregrine_unfriendly")
is_pff("/tmp/peregrine_unfriendly")
is_pff("/data/peregrine_unfriendly")
}
\seealso{
use \link{get_pff_tempfile}
to get a Peregrine-friendly filename
}
\author{
Richèl J.C. Bilderbeek
}
