% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pff_tempfile.R
\name{get_pff_tempfile}
\alias{get_pff_tempfile}
\title{Get a Peregrine-friendly temporary filename}
\usage{
get_pff_tempfile(
  pattern = "pff_",
  pff_tmpdir = get_pff_tempdir(),
  fileext = ""
)
}
\arguments{
\item{pattern}{see \link{tempfile}}

\item{pff_tmpdir}{Peregrine-friendly temporary directory name,
see \link{get_pff_tempdir}}

\item{fileext}{see \link{tempfile}}
}
\description{
Get a Peregrine-friendly temporary filename
}
\examples{
is_pff(get_pff_tempdir())
is_pff(get_pff_tempfile())
is_pff("/peregrine_unfriendly")
is_pff("/tmp/peregrine_unfriendly")
is_pff("/data/peregrine_unfriendly")
}
\seealso{
use \link{get_pff_tempdir}
to get a Peregrine-friendly folder name
}
\author{
Richèl J.C. Bilderbeek
}
