% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_on_peregrine.R
\name{is_on_peregrine}
\alias{is_on_peregrine}
\title{Is this code running on Peregrine?}
\usage{
is_on_peregrine()
}
\value{
TRUE if yes, FALSE if no
}
\description{
Is this code running on Peregrine?
}
\author{
Richèl J.C. Bilderbeek
}
