% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_on_peregrine_login_node.R
\name{is_on_peregrine_login_node}
\alias{is_on_peregrine_login_node}
\title{Is this code run on the Peregrine login node?}
\usage{
is_on_peregrine_login_node()
}
\value{
TRUe if yes, FALSE if np
}
\description{
Is this code run on the Peregrine login node?
}
\author{
Richèl J.C. Bilderbeek
}
