% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_on_peregrine_worker_node.R
\name{is_on_peregrine_worker_node}
\alias{is_on_peregrine_worker_node}
\title{Is this code run on a Peregrine worker node?}
\usage{
is_on_peregrine_worker_node()
}
\value{
TRUe if yes, FALSE if np
}
\description{
Is this code run on a Peregrine worker node?
}
\author{
Richèl J.C. Bilderbeek
}
