% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pff_beast2_options.R
\name{is_pff_beast2_options}
\alias{is_pff_beast2_options}
\title{Are the BEAST2 options PFF?}
\usage{
is_pff_beast2_options(beast2_options)
}
\arguments{
\item{beast2_options}{BEAST2 options,
as can be created by \link[beastier]{create_beast2_options}}
}
\description{
Are the BEAST2 options PFF?
}
