% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pff_experiment.R
\name{is_pff_experiment}
\alias{is_pff_experiment}
\title{Determine if an experiment has Peregrine friendly filenames}
\usage{
is_pff_experiment(experiment)
}
\arguments{
\item{experiment}{a \code{pirouette} experiment,
as created by \link[pirouette]{create_experiment}}
}
\description{
Determine if an experiment has Peregrine friendly filenames
}
