% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_strs_to_n_secs.R
\name{time_strs_to_n_secs}
\alias{time_strs_to_n_secs}
\title{Convert each time string to the number of seconds}
\usage{
time_strs_to_n_secs(strs)
}
\arguments{
\item{strs}{the time strings as text, for example \code{"12:34:56"}}
}
\description{
Convert each time string to the number of seconds
}
\examples{
time_strs <- c("00:00:01", "01:02:03", "1-02:03:04")

# Reduplicated plural: a vector of number of seconds
time_strs_to_n_secs(time_strs)
}
