% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pff.R
\name{to_pff}
\alias{to_pff}
\title{Convert a filename or folder to it Peregrine-friendly equivalent}
\usage{
to_pff(filename)
}
\arguments{
\item{filename}{the file's name, without the path}
}
\description{
Convert a filename or folder to it Peregrine-friendly equivalent
}
\examples{
# Filenames
filename <- "/peregrine_unfriendly.csv"
pff <- to_pff(filename)

filename <- "/tmp/peregrine_unfriendly.csv"
pff <- to_pff(filename)

filename <- "/data/peregrine_unfriendly.csv"
pff <- to_pff(filename)

# Folder names
foldername <- "/peregrine_unfriendly"
pff <- to_pff(foldername)

foldername <- "/tmp/peregrine_unfriendly"
pff <- to_pff(foldername)

foldername <- "/data/peregrine_unfriendly"
pff <- to_pff(foldername)
}
