% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pff_alignment_params.R
\name{to_pff_alignment_params}
\alias{to_pff_alignment_params}
\title{Convert all the alignment_params with Peregrine-unfriendly
filenames to ones that do}
\usage{
to_pff_alignment_params(alignment_params)
}
\arguments{
\item{alignment_params}{parameters for creating an alignment,
as can be created by \code{\link[pirouette]{create_alignment_params}}}
}
\description{
Convert all the alignment_params with Peregrine-unfriendly
filenames to ones that do
}
