% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pff_beast2_options.R
\name{to_pff_beast2_options}
\alias{to_pff_beast2_options}
\title{Converts a BEAST2 options to have PFF.
If all filenames are Peregrine-friendly,
this function does not alter the input}
\usage{
to_pff_beast2_options(beast2_options)
}
\description{
Converts a BEAST2 options to have PFF.
If all filenames are Peregrine-friendly,
this function does not alter the input
}
