% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pff_experiment.R
\name{to_pff_experiment}
\alias{to_pff_experiment}
\title{Convert an experiment to one with PFF.}
\usage{
to_pff_experiment(experiment)
}
\arguments{
\item{experiment}{a \link{pirouette} experiment,
as can be created by \link{create_experiment}}
}
\description{
Convert an experiment to one with PFF.
}
