% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pff_mcmc.R
\name{to_pff_mcmc}
\alias{to_pff_mcmc}
\title{Converts a mcmc to have PFF.
If all filenames are Peregrine-friendly,
this function does not alter the input}
\usage{
to_pff_mcmc(mcmc)
}
\arguments{
\item{mcmc}{an MCMC,
as created by \link[beautier]{create_mcmc} or by
\link{create_pff_mcmc} for automatic peregrine friendly filenames check}
}
\description{
Converts a mcmc to have PFF.
If all filenames are Peregrine-friendly,
this function does not alter the input
}
