% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pff_pir_params.R
\name{to_pff_pir_params}
\alias{to_pff_pir_params}
\title{Convert a \code{pir_params} to a \code{pir_params}
with Peregrine friendly filenames}
\usage{
to_pff_pir_params(pir_params)
}
\arguments{
\item{pir_params}{\code{pirouette} parameters,
as created by \link[pirouette]{create_pir_params}}
}
\description{
Convert a \code{pir_params} to a \code{pir_params}
with Peregrine friendly filenames
}
