% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pff_twinning_params.R
\name{to_pff_twinning_params}
\alias{to_pff_twinning_params}
\title{Convert a \code{twinning_params} to one with
Peregrine-friendly filenames}
\usage{
to_pff_twinning_params(twinning_params)
}
\arguments{
\item{twinning_params}{parameters for creating a twin tree,
as can be created by \code{\link[pirouette]{create_twinning_params}}}
}
\description{
Convert a \code{twinning_params} to one with
Peregrine-friendly filenames
}
