% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pff_inference_model.R
\name{check_pff_inference_model}
\alias{check_pff_inference_model}
\title{Check if the inference model is Peregrine friendly}
\usage{
check_pff_inference_model(inference_model)
}
\arguments{
\item{inference_model}{an Bayesian phylogenetic inference model,
as can be created by \code{\link{create_inference_model}}.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.}
}
\description{
Will \link{stop} if not
}
