% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pff_inference_model.R
\name{check_pff_inference_model}
\alias{check_pff_inference_model}
\title{Check if the inference model is Peregrine friendly}
\usage{
check_pff_inference_model(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link[beautier]{create_inference_model} to create an inference model.
Use \link[beautier]{check_inference_model} to check if  an inference model is valid.
Use \link[beautier]{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\description{
Will \link{stop} if not
}
