% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pff_alignment_params.R
\name{create_pff_alignment_params}
\alias{create_pff_alignment_params}
\title{Create alignment parameters with PFF}
\usage{

  create_pff_alignment_params(root_sequence = pirouette::create_alignment_params()$root_sequence,
  mutation_rate = pirouette::create_alignment_params()$mutation_rate,
  site_model = pirouette::create_alignment_params()$site_model,
  clock_model = pirouette::create_alignment_params()$clock_model,
  rng_seed = pirouette::create_alignment_params()$rng_seed,
  fasta_filename = get_pff_tempfile(pattern = "alignment_", fileext =
  ".fasta"))
}
\description{
Create alignment parameters with PFF
}
