% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pff_beast2_options.R
\name{create_pff_beast2_options}
\alias{create_pff_beast2_options}
\title{Create BEAST2 options that has Peregrine friendly filenames}
\usage{

  create_pff_beast2_options(input_filename = peregrine::get_pff_tempfile(),
  output_state_filename = peregrine::get_pff_tempfile())
}
\arguments{
\item{input_filename}{BEAST2 input file name,
as used in \link[beastier]{create_beast2_options}}

\item{output_state_filename}{BEAST2 ouput \code{.xml.state} file name,
as used in \link[beastier]{create_beast2_options}}
}
\description{
Create BEAST2 options that has Peregrine friendly filenames
}
\author{
Richel J.C. Bilderbeek
}
