% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pff_mcmc.R
\name{create_pff_mcmc}
\alias{create_pff_mcmc}
\title{Create mcmc that has Peregrine friendly filenames}
\usage{
create_pff_mcmc(chain_length, store_every,
  tracelog = beautier::create_tracelog(filename =
  peregrine::get_pff_tempfile()),
  treelog = beautier::create_treelog(filename =
  peregrine::get_pff_tempfile()))
}
\arguments{
\item{chain_length}{something}
}
\description{
Create mcmc that has Peregrine friendly filenames
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
