% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pff_pir_params.R
\name{create_pff_pir_params}
\alias{create_pff_pir_params}
\title{Create a \code{pir_params} with PFFs}
\usage{
create_pff_pir_params(
  alignment_params,
  twinning_params = NA,
  experiments = list(pirouette::create_experiment()),
  error_measure_params = pirouette::create_error_measure_params(),
  evidence_filename = NA,
  verbose = FALSE
)
}
\arguments{
\item{alignment_params}{parameters for creating an alignment,
as can be created by \code{\link[pirouette]{create_alignment_params}}}

\item{twinning_params}{parameters for creating a twin tree,
as can be created by \code{\link[pirouette]{create_twinning_params}}}

\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}. If more than one experiment
is provided and a "generative" experiment is part of them, the "generative"
one has to be the first in the list.}

\item{error_measure_params}{parameters to set how the error
between given tree and inferred trees in measure,
as can be created by \code{\link[pirouette]{create_error_measure_params}}}

\item{evidence_filename}{filename to store the estimated
evidences (aka marginal likelihoods)}

\item{verbose}{give more output}
}
\description{
Create a \code{pir_params} with PFFs
}
\author{
Richel J.C. Bilderbeek
}
