% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pff_tempdir.R
\name{get_pff_tempdir}
\alias{get_pff_tempdir}
\title{Get a Peregrine-friendly temporary directory}
\usage{
get_pff_tempdir()
}
\description{
Get a Peregrine-friendly temporary directory
}
\examples{
library(testthat)

expect_true(is_pff(get_pff_tempdir()))
expect_true(is_pff(get_pff_tempfile()))
expect_false(is_pff("/local/tmp/peregrine_unfriendly"))
expect_false(is_pff("/tmp/peregrine_unfriendly"))
}
\seealso{
use \link{get_pff_tempfile}
to get a Peregrine-friendly filename
}
\author{
Richel J.C. Bilderbeek
}
