% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pff_tempfile.R
\name{get_pff_tempfile}
\alias{get_pff_tempfile}
\title{Get a Peregrine-friendly temporary filename}
\usage{
get_pff_tempfile(pattern = "razzo_file_",
  pff_tmpdir = get_pff_tempdir(), fileext = "")
}
\arguments{
\item{pattern}{see \link{tempfile}}

\item{pff_tmpdir}{Peregrine-friendly temporary directory name,
see \link{get_pff_tempdir}}

\item{fileext}{see \link{tempfile}}
}
\description{
Get a Peregrine-friendly temporary filename
}
\examples{
library(testthat)

expect_true(is_pff(get_pff_tempdir()))
expect_true(is_pff(get_pff_tempfile()))
expect_false(is_pff("/local/tmp/peregrine_unfriendly"))
expect_false(is_pff("/tmp/peregrine_unfriendly"))
}
\seealso{
use \link{get_pff_tempdir}
to get a Peregrine-friendly folder name
}
\author{
Richel J.C. Bilderbeek
}
