% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pff.R
\name{is_pff}
\alias{is_pff}
\title{Is the filename a Peregrine-friendly filename?}
\usage{
is_pff(filename)
}
\arguments{
\item{filename}{name of the file}
}
\description{
Is the filename a Peregrine-friendly filename?
}
\examples{
library(testthat)

expect_true(is_pff(get_pff_tempdir()))
expect_true(is_pff(get_pff_tempfile()))
expect_false(is_pff("/local/tmp/peregrine_unfriendly"))
expect_false(is_pff("/tmp/peregrine_unfriendly"))
}
\seealso{
see \link{get_pff_tempdir} and \link{get_pff_tempfile}
to get a Peregrine-friendly temporary folder and filename
}
\author{
Richel J.C. Bilderbeek
}
