% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_strs_to_n_secs.R
\name{time_strs_to_n_secs}
\alias{time_strs_to_n_secs}
\title{Convert each time string to the number of seconds}
\usage{
time_strs_to_n_secs(strs)
}
\arguments{
\item{str}{the time string as text, for example \code{"12:34:56"}}
}
\description{
Convert each time string to the number of seconds
}
\examples{
library(testthat)
time_strs <- c("00:00:01", "01:02:03", "1-02:03:04")

# Reduplicated plural: a vector of number of seconds
n_secses <- time_strs_to_n_secs(time_strs)

expect_true(all(n_secses > 0))
}
