% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pff.R
\name{to_pff}
\alias{to_pff}
\title{Convert a filename or folder to it Peregrine-friendly equivalent}
\usage{
to_pff(filename)
}
\arguments{
\item{filename}{the file's name, without the path}
}
\description{
Convert a filename or folder to it Peregrine-friendly equivalent
}
\examples{
library(testthat)

# Filenames

filename <- "/local/tmp/peregrine_unfriendly.csv"
expect_false(is_pff(filename))
pff <- to_pff(filename)
expect_true(is_pff(pff))
expect_equal(basename(filename), basename(pff))

filename <- "/tmp/peregrine_unfriendly.csv"
expect_false(is_pff(filename))
pff <- to_pff(filename)
expect_true(is_pff(pff))
expect_equal(basename(filename), basename(pff))

# Folder names

foldername <- "/local/tmp/peregrine_unfriendly"
expect_false(is_pff(foldername))
pff <- to_pff(filename)
expect_true(is_pff(pff))
expect_equal(basename(foldername), basename(pff))

foldername <- "/tmp/peregrine_unfriendly"
expect_false(is_pff(foldername))
pff <- to_pff(filename)
expect_true(is_pff(pff))
expect_equal(basename(foldername), basename(pff))
}
