% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pff_experiments.R
\name{to_pff_experiments}
\alias{to_pff_experiments}
\title{Convert all the experiments with Peregrine-unfriendly
filenames to ones that do}
\usage{
to_pff_experiments(experiments)
}
\description{
Convert all the experiments with Peregrine-unfriendly
filenames to ones that do
}
